+++ kong/runloop/handler.lua
@@ -39,12 +39,14 @@ local ngx          = ngx
 local var          = ngx.var
 local log          = ngx.log
 local exit         = ngx.exit
+local exec         = ngx.exec
 local null         = ngx.null
 local header       = ngx.header
 local timer_at     = ngx.timer.at
 local timer_every  = ngx.timer.every
 local subsystem    = ngx.config.subsystem
 local clear_header = ngx.req.clear_header
+local http_version = ngx.req.http_version
 local unpack       = unpack
 local escape       = require("kong.tools.uri").escape
 
@@ -1209,7 +1211,6 @@ return {
 
       ctx.workspace = match_t.route and match_t.route.ws_id
 
-      local http_version   = ngx.req.http_version()
       local host           = var.host
       local port           = tonumber(ctx.host_port, 10)
                           or tonumber(var.server_port, 10)
@@ -1283,13 +1284,13 @@ return {
         end
       end
 
+      local protocol_version = http_version()
       if protocols.grpc or protocols.grpcs then
         -- perf: branch usually not taken, don't cache var outside
         local content_type = var.content_type
 
         if content_type and sub(content_type, 1, #"application/grpc") == "application/grpc" then
-          http_version = ngx.req.http_version()
-          if http_version ~= 2 then
+          if protocol_version ~= 2 then
             -- mismatch: non-http/2 request matched grpc route
             return kong.response.exit(426, { message = "Please use HTTP2 protocol" }, {
               ["connection"] = "Upgrade",
@@ -1358,20 +1359,20 @@ return {
       -- to `grpc_pass`. After redirection, this function will return early
       if service and var.kong_proxy_mode == "http" then
         if service.protocol == "grpc" or service.protocol == "grpcs" then
-          return ngx.exec("@grpc")
+          return exec("@grpc")
         end
 
-        if http_version == 1.1 then
+        if protocol_version == 1.1 then
           if route.request_buffering == false then
             if route.response_buffering == false then
-              return ngx.exec("@unbuffered")
+              return exec("@unbuffered")
             end
 
-            return ngx.exec("@unbuffered_request")
+            return exec("@unbuffered_request")
           end
 
           if route.response_buffering == false then
-            return ngx.exec("@unbuffered_response")
+            return exec("@unbuffered_response")
           end
         end
       end
@@ -1427,9 +1428,8 @@ return {
 
       local ok, err = balancer.set_host_header(balancer_data, upstream_scheme, upstream_host)
       if not ok then
-        ngx.log(ngx.ERR, "failed to set balancer Host header: ", err)
-
-        return ngx.exit(500)
+        log(ERR, "failed to set balancer Host header: ", err)
+        return exit(500)
       end
 
       -- clear hop-by-hop request headers: